+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file. `Jayme` a
 - `path` variable has been renamed to `name` in `Repository` protocol declaration.
 - `ServerBackend` protocol has been renamed to `NSURLSessionBackend`.
 - `ServerBackendConfiguration` has been renamed to `NSURLSessionBackendConfiguration`.
+- `ServerBackendError` has been renamed to `JaymeError`.
 
 ---